feat: atualizar DTO do prestador para validar CPF com formato e adicionar upload de imagem, incluindo busca por token de autenticação

+++ backend/src/modules/prestador/prestador.service.ts
@@ -1,21 +1,40 @@
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreatePrestadorDTO } from './dto/prestador.dto';
 import { randomBytes } from 'crypto';
-import { BadRequestException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
+import { UploadService } from '../upload/upload.service';
 
+@Injectable()
 export class PrestadorService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(
+    private readonly prisma: PrismaService,
+    private readonly uploadService: UploadService,
+  ) {}
 
   async create({ nome, cpf, foto, telefone, descricao }: CreatePrestadorDTO) {
+    const existingPrestador = await this.prisma.prestadorDeServico.findUnique({
+      where: { cpf },
+    });
+
+    if (existingPrestador) {
+      throw new BadRequestException('CPF já cadastrado');
+    }
+
+    // Se houver foto e for necessário realizar o upload
+    if (foto) {
+      const uploadedUrl = await this.uploadService.uploadImage(
+        foto,
+        cpf,
+        'prestador',
+      );
+      foto = uploadedUrl; // Atualiza com a URL retornada
+    }
+
     try {
-      const existingCliente = await this.prisma.prestadorDeServico.findUnique({
-        where: { cpf },
-      });
-
-      if (existingCliente) {
-        throw new BadRequestException('CPF já cadastrado');
-      }
-
       const prestador = await this.prisma.prestadorDeServico.create({
         data: {
           nome,
@@ -31,7 +50,10 @@ export class PrestadorService {
 
       await this.prisma.prestadorDeServico.update({
         where: { id: prestador.id },
-        data: { auth_token: authToken },
+        data: {
+          auth_token: authToken,
+          expire_date: new Date(new Date().setDate(new Date().getDate() + 10)),
+        },
       });
 
       return { auth_token: authToken };
@@ -55,6 +77,23 @@ export class PrestadorService {
   }
 
   async remove(id: string): Promise<any> {
+    const prestador = await this.prisma.prestadorDeServico.findUnique({
+      where: { id },
+    });
+    if (!prestador) {
+      throw new NotFoundException('Prestador não encontrado');
+    }
+
+    // Se existir uma foto, utiliza o UploadService para removê-la do Cloudinary
+    if (prestador.foto) {
+      await this.uploadService.removeImage(prestador.foto);
+    }
+
     return this.prisma.prestadorDeServico.delete({ where: { id } });
+  }
+  async findByAuthToken(authToken: string) {
+    return this.prisma.prestadorDeServico.findUnique({
+      where: { auth_token: authToken },
+    });
   }
 }
